thermalrenderer/requirements.txt
influxdb-client
paramiko
pillow
numpy
matplotlib
